enhanced_requirements.txt
Flask==2.3.3
flask-cors==4.0.0
PyPDF2==3.0.1
Werkzeug==2.3.7
Pillow==10.0.0
pytesseract==0.3.10
pdf2image==1.16.3
